+++ Runtime/Unity.Serialization/Json/Views/SerializedObjectView.cs
@@ -119,7 +119,6 @@ namespace Unity.Serialization.Json
         /// </summary>
         /// <param name="name">The key of the value to get.</param>
         /// <exception cref="KeyNotFoundException">The key does not exist in the collection.</exception>
-        [NotBurstCompatible] 
         public SerializedValueView this[string name] => GetValue(name);
         
         /// <summary>
@@ -148,7 +147,6 @@ namespace Unity.Serialization.Json
         /// </summary>
         /// <param name="name">The key of the member to get.</param>
         /// <returns>Returns the member associated with the specified key.</returns>
-        [NotBurstCompatible]
         public SerializedMemberView GetMember(string name)
         {
             if (!TryGetMember(name, out var member))
@@ -163,7 +161,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the member to get.</param>
         /// <param name="member">When this method returns, contains the member associated with the specified key, if the key is found; otherwise, the default value.</param>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key; otherwise, false.</returns>
-        [NotBurstCompatible]
         public bool TryGetMember(string name, out SerializedMemberView member)
         {
             foreach (var m in this)
@@ -186,7 +183,6 @@ namespace Unity.Serialization.Json
         /// </summary>
         /// <param name="name">The key of the value to get.</param>
         /// <returns>Returns the value associated with the specified key.</returns>
-        [NotBurstCompatible]
         public SerializedValueView GetValue(string name)
         {
             if (!TryGetValue(name, out var value))
@@ -201,7 +197,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the value to get.</param>
         /// <param name="value">When this method returns, contains the value associated with the specified key, if the key is found; otherwise, the default value.</param>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key; otherwise, false.</returns>
-        [NotBurstCompatible]
         public bool TryGetValue(string name, out SerializedValueView value)
         {
             foreach (var m in this)
@@ -225,7 +220,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the value to get.</param>
         /// <param name="value">When this method returns, contains the value associated with the specified key and type, if the key is found and the type matches; otherwise, the default value.</param>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key and type; otherwise, false.</returns>
-        [NotBurstCompatible]
         public bool TryGetValueAsString(string name, out string value)
         {
             value = default;
@@ -246,7 +240,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the value to get.</param>
         /// <param name="value">When this method returns, contains the value associated with the specified key and type, if the key is found and the type matches; otherwise, the default value.</param>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key and type; otherwise, false.</returns>
-        [NotBurstCompatible]
         public bool TryGetValueAsInt64(string name, out long value)
         {
             value = default;
@@ -268,7 +261,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the value to get.</param>
         /// <param name="value">When this method returns, contains the value associated with the specified key and type, if the key is found and the type matches; otherwise, the default value.</param>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key and type; otherwise, false.</returns>
-        [NotBurstCompatible]
         public bool TryGetValueAsUInt64(string name, out ulong value)
         {
             value = default;
@@ -290,7 +282,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the value to get.</param>
         /// <param name="value">When this method returns, contains the value associated with the specified key and type, if the key is found and the type matches; otherwise, the default value.</param>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key and type; otherwise, false.</returns>
-        [NotBurstCompatible]
         public bool TryGetValueAsFloat(string name, out float value)
         {
             value = default;
@@ -312,7 +303,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the value to get.</param>
         /// <param name="value">When this method returns, contains the value associated with the specified key and type, if the key is found and the type matches; otherwise, the default value.</param>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key and type; otherwise, false.</returns>
-        [NotBurstCompatible]
         public bool TryGetValueAsDouble(string name, out double value)
         {
             value = default;
@@ -334,7 +324,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the value to get.</param>
         /// <param name="value">When this method returns, contains the value associated with the specified key and type, if the key is found and the type matches; otherwise, the default value.</param>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key and type; otherwise, false.</returns>
-        [NotBurstCompatible]
         public bool TryGetValueAsBoolean(string name, out bool value)
         {
             value = default;
@@ -356,7 +345,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the value to get.</param>
         /// <typeparam name="T">The fixed string type.</typeparam>
         /// <returns>Returns the value associated with the specified key.</returns>
-        [BurstCompatible(GenericTypeArguments = new[] { typeof(FixedString128Bytes) })]
         public SerializedValueView GetValue<T>(in T name) where T : unmanaged, INativeList<byte>, IUTF8Bytes
         {
             if (!TryGetValue(name, out var value))
@@ -372,7 +360,6 @@ namespace Unity.Serialization.Json
         /// <param name="value">When this method returns, contains the value associated with the specified key, if the key is found; otherwise, the default value.</param>
         /// <typeparam name="T">The fixed string type.</typeparam>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key; otherwise, false.</returns>
-        [BurstCompatible(GenericTypeArguments = new[] { typeof(FixedString128Bytes) })]
         public bool TryGetValue<T>(in T name, out SerializedValueView value) where T : unmanaged, INativeList<byte>, IUTF8Bytes
         {
             foreach (var m in this)
@@ -396,7 +383,6 @@ namespace Unity.Serialization.Json
         /// <param name="name">The key of the member to get.</param>
         /// <typeparam name="T">The fixed string type.</typeparam>
         /// <returns>Returns the member associated with the specified key.</returns>
-        [BurstCompatible(GenericTypeArguments = new[] { typeof(FixedString128Bytes) })]
         public SerializedMemberView GetMember<T>(in T name) where T : unmanaged, INativeList<byte>, IUTF8Bytes
         {
             if (!TryGetMember(name, out var member))
@@ -412,7 +398,6 @@ namespace Unity.Serialization.Json
         /// <typeparam name="T">The fixed string type.</typeparam>
         /// <param name="member">When this method returns, contains the member associated with the specified key, if the key is found; otherwise, the default value.</param>
         /// <returns>true if the <see cref="SerializedObjectView"/> contains an element with the specified key; otherwise, false.</returns>
-        [BurstCompatible(GenericTypeArguments = new[] { typeof(FixedString128Bytes) })]
         public bool TryGetMember<T>(in T name, out SerializedMemberView member) where T : unmanaged, INativeList<byte>, IUTF8Bytes
         {
             foreach (var m in this)