+++ CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2022-09-29
+### Changed
+* Removed `[NotBurstCompatible]` and `[BurstCompatible]` attributes to support newer versions of collections.
+
 ## [2.1.0] - 2022-08-12
 ### Added
 * Added `SerializedObjectReader.SetSource` method to allow re-use of the same instance.
 * Added `FixedString` support for all `SerializedViews`.
 
 ### Changed
-* Updated `com.unity.properties` to version `2.1.0-exp.1`.
+* Updated `com.unity.properties` to version `2.1.0-exp.2`.
 * Updated `com.unity.collections` to version `1.4.0`.
 * Updated `com.unity.burst` to version `1.7.2`.
 * Updated `com.unity.jobs` to version `0.70.0-preview.7`.